requirements.txt
click>=8.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.7.0
pyperclip>=1.8.2
requests>=2.31.0
